requirements.txt
appdirs==1.4.3
Django==1.10.6
django-crispy-forms==1.6.1
django-filter==1.0.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.6.3
feedparser==5.2.1
markdown2==2.3.3
mysqlclient==1.3.10
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
